requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyjwt==2.8.0
werkzeug==3.0.1
requests==2.31.0
pycups==2.0.1
gunicorn==21.2.0 